refactor: correcciones pull request
