+++ libs/shared-validation/src/lib/enums/user.ts
@@ -3,5 +3,3 @@ export enum UserRolesEnum {
   USER = 'USER',
   AUCTION_MANAGER = 'AUCTION_MANAGER',
 }
-
-// cambiado a mayusculas, ya que daba error
